third_party/blink/web_tests/platform/generic/fast/table/col-width-span-expand-expected.txt

Two rows of cells should look identical.

left	middle	right
left	middle		right

This tests Bug 14858: <col> width ignored when not tied to a single cell.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS window.getComputedStyle(testCell).width is '250px'
PASS window.getComputedStyle(testTable).width is '500px'
PASS successfullyParsed is true

TEST COMPLETE
